Move some logic from graphql middlewares to context setup
