+++ saleor/graphql/core/tests/test_context.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 from django.urls import reverse
 
-from ...middleware import app_middleware
+from ...context import set_app_on_context
 
 
 def test_app_middleware_accepts_app_requests(app, rf):
@@ -13,7 +13,7 @@ def test_app_middleware_accepts_app_requests(app, rf):
     request.META = {"HTTP_AUTHORIZATION": f"Bearer {token}"}
 
     # when
-    app_middleware(lambda root, info: info.context, Mock(), Mock(context=request))
+    set_app_on_context(request)
 
     # then
     assert request.app == app
@@ -26,7 +26,7 @@ def test_app_middleware_accepts_saleors_header(app, rf):
     request.META = {"HTTP_AUTHORIZATION_BEARER": f"{token}"}
 
     # when
-    app_middleware(lambda root, info: info.context, Mock(), Mock(context=request))
+    set_app_on_context(request)
 
     # then
     assert request.app == app
@@ -38,7 +38,7 @@ def test_app_middleware_skips_when_token_length_is_different_than_30(app, rf):
     request.META = {"HTTP_AUTHORIZATION_BEARER": "a" * 31}
 
     # when
-    app_middleware(lambda root, info: info.context, Mock(), Mock(context=request))
+    set_app_on_context(request)
 
     # then
     assert not request.app
